+++ neqo-common/src/datagram.rs
@@ -11,34 +11,18 @@ use std::{
 
 use crate::{hex_with_len, IpTos};
 
-/// The meta data associated with a UDP datagram.
 #[derive(Clone, PartialEq, Eq)]
-pub struct DatagramMetaData {
+pub struct Datagram<D = Vec<u8>> {
     src: SocketAddr,
     dst: SocketAddr,
     tos: IpTos,
-    len: usize,
+    d: D,
 }
 
-impl DatagramMetaData {
+impl<D> Datagram<D> {
     #[must_use]
-    pub const fn new(src: SocketAddr, dst: SocketAddr, tos: IpTos, len: usize) -> Self {
-        Self { src, dst, tos, len }
-    }
-
-    #[must_use]
-    pub const fn len(&self) -> usize {
-        self.len
-    }
-
-    #[must_use]
-    pub const fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    #[must_use]
-    pub const fn tos(&self) -> IpTos {
-        self.tos
+    pub const fn source(&self) -> SocketAddr {
+        self.src
     }
 
     #[must_use]
@@ -47,46 +31,12 @@ impl DatagramMetaData {
     }
 
     #[must_use]
-    pub const fn source(&self) -> SocketAddr {
-        self.src
-    }
-}
-
-/// Whether the given datagram matches this meta data.
-impl PartialEq<Datagram> for DatagramMetaData {
-    fn eq(&self, other: &Datagram) -> bool {
-        self.dst == other.destination() && self.len == other.len() && self.tos == other.tos()
-    }
-}
-
-#[derive(Clone, PartialEq, Eq)]
-pub struct Datagram<D = Vec<u8>> {
-    meta: DatagramMetaData,
-    d: D,
-}
-
-impl<D> Datagram<D> {
-    #[must_use]
-    pub const fn source(&self) -> SocketAddr {
-        self.meta.source()
-    }
-
-    #[must_use]
-    pub const fn destination(&self) -> SocketAddr {
-        self.meta.destination()
-    }
-
-    #[must_use]
     pub const fn tos(&self) -> IpTos {
-        self.meta.tos()
+        self.tos
     }
 
     pub fn set_tos(&mut self, tos: IpTos) {
-        self.meta.tos = tos;
-    }
-
-    pub const fn meta(&self) -> &DatagramMetaData {
-        &self.meta
+        self.tos = tos;
     }
 }
 
@@ -108,15 +58,11 @@ impl<D: AsMut<[u8]> + AsRef<[u8]>> AsMut<[u8]> for Datagram<D> {
 
 impl Datagram<Vec<u8>> {
     pub fn new<V: Into<Vec<u8>>>(src: SocketAddr, dst: SocketAddr, tos: IpTos, d: V) -> Self {
-        let d = d.into();
         Self {
-            meta: DatagramMetaData {
-                src,
-                dst,
-                len: d.len(),
-                tos,
-            },
-            d,
+            src,
+            dst,
+            tos,
+            d: d.into(),
         }
     }
 }
@@ -139,9 +85,9 @@ impl<D: AsRef<[u8]>> std::fmt::Debug for Datagram<D> {
         write!(
             f,
             "Datagram {:?} {:?}->{:?}: {}",
-            self.meta.tos(),
-            self.meta.source(),
-            self.meta.destination(),
+            self.tos,
+            self.src,
+            self.dst,
             hex_with_len(&self.d)
         )
     }
@@ -150,21 +96,15 @@ impl<D: AsRef<[u8]>> std::fmt::Debug for Datagram<D> {
 impl<'a> Datagram<&'a mut [u8]> {
     #[must_use]
     pub fn from_slice(src: SocketAddr, dst: SocketAddr, tos: IpTos, d: &'a mut [u8]) -> Self {
-        Self {
-            meta: DatagramMetaData {
-                src,
-                dst,
-                len: d.len(),
-                tos,
-            },
-            d,
-        }
+        Self { src, dst, tos, d }
     }
 
     #[must_use]
     pub fn to_owned(&self) -> Datagram {
         Datagram {
-            meta: self.meta.clone(),
+            src: self.src,
+            dst: self.dst,
+            tos: self.tos,
             d: self.d.to_vec(),
         }
     }
